backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.6
httpx==0.26.0
celery==5.3.6
redis==5.0.1
stripe==8.2.0
cryptography==42.0.0
python-dotenv==1.0.0
email-validator>=2.0
alembic==1.13.1

